requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
cycler==0.11.0
decorator==5.1.1
docker-pycreds==0.4.0
executing==0.8.3
filelock==3.6.0
fonttools==4.32.0
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.5.1
idna==3.3
imageio==2.16.2
ipdb==0.13.9
ipython==8.2.0
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.4.2
lmdb==1.3.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
MultiScaleDeformableAttention==1.0
networkx==2.8
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095647912/work
opencv-python==4.5.5.64
packaging==21.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1648857110829/work
prettytable==3.2.0
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycocotools==2.0.4
Pygments==2.11.2
pyparsing==3.0.8
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.3.15
requests==2.27.1
sacremoses==0.0.49
scikit-image==0.19.2
scipy==1.8.0
sentry-sdk==1.5.9
setproctitle==1.2.2
shortuuid==1.0.8
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
stack-data==0.2.0
tifffile==2022.4.8
tokenizers==0.11.6
toml==0.10.2
torch==1.5.1
torchvision==0.6.0a0+35d732a
tqdm==4.64.0
traitlets==5.1.1
transformers==4.18.0
typing_extensions==4.1.1
urllib3==1.26.9
wandb==0.12.14
wcwidth==0.2.5
